🎨 쿠폰 관련 이슈 수정

+++ src/main/java/com/war11/domain/order/dto/response/OrderResponse.java
@@ -11,7 +11,7 @@ public record OrderResponse (
     Long orderId,
     List<OrderProductResponse> orderProducts,
     Long totalPrice,
-    Integer discountedPrice,
+    Long discountedPrice,
     Long resultPrice,
     OrderStatus orderStatus,
     LocalDateTime createdAt,
